+++ packages/ad-replay/src/internal/vars.ts
