+++ src/replay/target-identity-node.ts
@@ -19,7 +19,7 @@ import {
   truncateToUtf8Bytes,
   TARGET_ANNOTATION_MAX_FIELD_BYTES,
 } from '@agent-device/ad-script';
-import type { LocalIdentity } from './target-identity.ts';
+import type { LocalIdentity } from '@agent-device/ad-replay';
 
 type IdentityTreeNode = Pick<RawSnapshotNode, 'type' | 'identifier' | 'label'>;
